+++ base-domain/src/main/java/base/configuration/DomainConfig.java
@@ -1,0 +1,14 @@
+package base.configuration;
+
+import base.domain.Domain;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+@Configuration
+@EnableTransactionManagement
+@Import(value = { JpaRepositoryConfig.class })
+@ComponentScan(basePackageClasses = Domain.class)
+public class DomainConfig {
+}
