+++ base-domain/src/main/java/base/domain/Domain.java
@@ -1,0 +1,4 @@
+package base.domain;
+
+public interface Domain {
+}
