+++ base-domain/src/main/java/base/domain/member/MemberRepository.java
@@ -1,0 +1,12 @@
+package base.domain.member;
+
+import base.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+	Optional<Member> findMemberByMemberEmail(String memberEmail);
+}
