+++ base-domain/src/main/resources/db/migration/V1__CreateTable_Member.sql
