+++ base-domain/src/main/resources/db/migration/initialize.sql
