+++ base-interface/build.gradle
@@ -1,0 +1,11 @@
+dependencies {
+    implementation (project(':base-domain'))
+
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    compile group: 'io.springfox', name: 'springfox-swagger2', version: '2.9.2'
+    compile group: 'io.springfox', name: 'springfox-swagger-ui', version: '2.9.2'
+
+    testImplementation 'org.springframework.security:spring-security-test'
+}
