+++ base-interface/src/main/java/base/application/Application.java
@@ -1,0 +1,4 @@
+package base.application;
+
+public interface Application {
+}
