+++ base-interface/src/main/java/base/interfaces/GlobalExceptionHandler.java
@@ -1,0 +1,16 @@
+package base.interfaces;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@Slf4j
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+	@ExceptionHandler(Exception.class)
+	public String global(Exception e) {
+		log.error("Exception : {}", e.getMessage(), e);
+		return "Exception";
+	}
+}
