+++ base-interface/src/main/java/base/interfaces/HelloController.java
@@ -1,0 +1,13 @@
+package base.interfaces;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloController {
+
+	@GetMapping("/hello")
+	public String hello() {
+		return "Hello Base Project";
+	}
+}
