+++ base-interface/src/main/java/base/interfaces/Interfaces.java
@@ -1,0 +1,4 @@
+package base.interfaces;
+
+public interface Interfaces {
+}
