+++ build.gradle
@@ -1,57 +1,59 @@
 plugins {
-	id 'org.springframework.boot' version '2.2.0.RELEASE'
-	id 'io.spring.dependency-management' version '1.0.8.RELEASE'
 	id 'java'
 	id 'groovy'
+	id 'org.springframework.boot' version '2.2.0.RELEASE'
+	id 'io.spring.dependency-management' version '1.0.8.RELEASE'
 	id "org.flywaydb.flyway" version "6.0.6"
 }
 
-apply from: 'flyway.gradle'
-
-group = 'com.study'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
-
-configurations {
-	developmentOnly
-	runtimeClasspath {
-		extendsFrom developmentOnly
+allprojects {
+	repositories {
+		jcenter()
+		mavenCentral()
+		maven {
+			url "https://plugins.gradle.org/m2/"
+		}
 	}
-	compileOnly {
-		extendsFrom annotationProcessor
-	}
+
+	apply plugin: 'java'
+	apply plugin: 'groovy'
+	apply plugin: 'idea'
+	apply plugin: 'eclipse'
+	apply plugin: 'org.springframework.boot'
+	apply plugin: 'io.spring.dependency-management'
 }
 
-repositories {
-	mavenCentral()
+bootJar {
+	mainClassName = 'base.BaseApplication'
 }
 
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	implementation 'org.springframework.boot:spring-boot-starter-security'
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'org.flywaydb:flyway-core'
+subprojects {
+	group = 'com.study'
+	version = '0.0.1-SNAPSHOT'
+	sourceCompatibility = '1.8'
 
-	compile group: 'mysql', name: 'mysql-connector-java', version: '8.0.18'
-	compile group: 'com.google.guava', name: 'guava', version: '28.1-jre'
-	compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.9'
-	compile group: 'io.springfox', name: 'springfox-swagger2', version: '2.9.2'
-	compile group: 'io.springfox', name: 'springfox-swagger-ui', version: '2.9.2'
+	dependencies {
+		implementation 'org.springframework.boot:spring-boot-starter'
 
-	compileOnly 'org.projectlombok:lombok'
-	developmentOnly 'org.springframework.boot:spring-boot-devtools'
+		compile group: 'com.google.guava', name: 'guava', version: '28.1-jre'
+		compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.9'
 
-	annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
-	annotationProcessor 'org.projectlombok:lombok'
+		compileOnly 'org.projectlombok:lombok'
 
-	testImplementation('org.springframework.boot:spring-boot-starter-test') {
-		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+		annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
+		annotationProcessor 'org.projectlombok:lombok'
+
+		testImplementation('org.springframework.boot:spring-boot-starter-test') {
+			exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+		}
+		testCompile('org.springframework.boot:spring-boot-starter-test')
+		testCompile group: 'org.spockframework', name: 'spock-spring', version: '1.3-groovy-2.5'
+		testCompile group: 'io.github.benas', name: 'random-beans-randomizers', version: '3.9.0'
 	}
-	testImplementation 'org.springframework.security:spring-security-test'
-	testCompile group: 'org.spockframework', name: 'spock-spring', version: '1.3-groovy-2.5'
-	testCompile group: 'io.github.benas', name: 'random-beans-randomizers', version: '3.9.0'
-}
 
-test {
-	useJUnitPlatform()
+	dependencyManagement {
+		imports {
+			mavenBom org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
+		}
+	}
 }
