--- flyway.gradle
@@ -1,10 +1,0 @@
-// MYSQL 5.7 이상부터 지원.
-flyway {
-	url = 'jdbc:mysql://localhost:3306/base?useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC'
-	driver = "com.mysql.cj.jdbc.Driver"
-	user = 'study'
-	password = 'hard'
-	schemas = ['base']
-	table = 'flyway_migration_history'
-	encoding = 'UTF8'
-}
