--- src/main/java/com/study/base/application/Application.java
@@ -1,4 +1,0 @@
-package com.study.base.application;
-
-public interface Application {
-}
