--- src/main/java/com/study/base/configration/ApplicationConfig.java
@@ -1,10 +1,0 @@
-package com.study.base.configration;
-
-import com.study.base.application.Application;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(basePackageClasses = Application.class)
-public class ApplicationConfig {
-}
