--- src/main/java/com/study/base/configration/InfraConfig.java
@@ -1,10 +1,0 @@
-package com.study.base.configration;
-
-import com.study.base.infra.Infra;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(basePackageClasses = Infra.class)
-public class InfraConfig {
-}
