--- src/main/java/com/study/base/domain/Domain.java
@@ -1,4 +1,0 @@
-package com.study.base.domain;
-
-public interface Domain {
-}
