--- src/main/java/com/study/base/domain/member/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.study.base.domain.member;
-
-import com.study.base.domain.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-	Optional<Member> findMemberByMemberEmail(String memberEmail);
-}
