--- src/main/java/com/study/base/domain/support/entity/CreatedAndModifiedEntity.java
@@ -1,22 +1,0 @@
-package com.study.base.domain.support.entity;
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(value = { AuditingEntityListener.class })
-public class CreatedAndModifiedEntity {
-
-	@CreatedDate
-	private LocalDateTime createdAt;
-
-	@LastModifiedDate
-	private LocalDateTime modifiedAt;
-}
