--- src/main/java/com/study/base/infra/Infra.java
@@ -1,4 +1,0 @@
-package com.study.base.infra;
-
-public interface Infra {
-}
