--- src/main/java/com/study/base/interfaces/GlobalExceptionHandler.java
@@ -1,16 +1,0 @@
-package com.study.base.interfaces;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@Slf4j
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-	@ExceptionHandler(Exception.class)
-	public String global(Exception e) {
-		log.error("Exception : {}", e.getMessage(), e);
-		return "Exception";
-	}
-}
