--- src/main/java/com/study/base/interfaces/HelloController.java
@@ -1,13 +1,0 @@
-package com.study.base.interfaces;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-
-	@GetMapping("/hello")
-	public String hello() {
-		return "Hello Base Project";
-	}
-}
