--- src/main/java/com/study/base/interfaces/Interfaces.java
@@ -1,4 +1,0 @@
-package com.study.base.interfaces;
-
-public interface Interfaces {
-}
