--- src/main/resources/application.yml
@@ -1,49 +1,0 @@
-# default environment
-spring:
-  flyway:
-    enabled: false
-  profiles:
-    active: local
-  groovy:
-    template:
-      check-template-location: false
-
-# local environment
----
-spring:
-  profiles: local
-  datasource:
-    url: jdbc:mysql://localhost:3306/base?useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC
-    username: study
-    password: hard
-  jpa:
-    show-sql: true
-    hibernate:
-      ddl-auto: validate
-      naming:
-        physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
-    open-in-view: false
-
-server:
-  port: 8080
-
-# production environment
----
-spring:
-  profiles: prod
-  datasource:
-    url: jdbc:mysql://localhost:3306/base?useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC
-    username: study
-    password: hard
-  jpa:
-    show-sql: false
-    hibernate:
-      ddl-auto: validate
-      naming:
-        physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
-    open-in-view: false
-
-server:
-  port: 8080
