--- src/test/java/com/study/base/BaseApplicationTests.java
@@ -1,15 +1,0 @@
-package com.study.base;
-
-import org.junit.Ignore;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@Ignore
-@SpringBootTest
-class BaseApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
